pyproject.toml
[project]
name = "deadlock-hero-ability-statistics-image-extractor"
version = "0.1.0"
description = "A cross-platform Python tool to automatically launch Deadlock and extract hero ability images and stat images on Windows and Linux."
authors = [
    {name = "Oskar Sterner", email = "oskar.lr.sterner@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "psutil>=5.9.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pynput>=1.7.6",
    "pyautogui>=0.9.54",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "requests>=2.28.0",
    "opencv-python>=4.8.0",
    "ultralytics"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "labelImg"
]

[project.scripts]
deadlock-extractor = "deadlock_hero_ability_statistics_image_extractor.main:main"
deadlock-extractor-web = "deadlock_hero_ability_statistics_image_extractor.web_app:run_web_app"
train-tooltip-detector = "deadlock_hero_ability_statistics_image_extractor.train_yolo:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/deadlock_hero_ability_statistics_image_extractor"]

[tool.ruff]
line-length = 88
target-version = "py39"
src = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']